+++ commands/create.go
@@ -2,6 +2,8 @@ package commands
 
 import (
 	"fmt"
+
+	"com.github/axblade/playwright/utils"
 )
 
 type CreateCommand struct {
@@ -9,8 +11,36 @@ type CreateCommand struct {
 }
 
 func (self *CreateCommand) Execute() (err error) {
-	fmt.Println("Execute of CreateCommand is called")
 	folders := self.SelectFolders()
-	fmt.Println(folders)
+
+	rolesPath, err := self.ReadRolesPath()
+	if err != nil {
+		return err
+	}
+
+	createPlaybookStructure(rolesPath, self.Command.Name, folders)
+
 	return nil
+}
+
+func createPlaybookStructure(rolesPath string, name string, folders []string) {
+	if string(rolesPath[len(rolesPath)-1]) != "/" {
+		rolesPath = utils.Concat(rolesPath, "/")
+	}
+
+	playbookPath := concat(rolesPath, name)
+
+	if string(playbookPath[len(playbookPath)-1]) != "/" {
+		playbookPath = utils.Concat(playbookPath, "/")
+	}
+
+	for _, folder := range folders {
+		folderPath := utils.Concat(playbookPath, folder)
+		os.MkdirAll(folderPath, 0755)
+
+		if folder != "files" && folder != "templates" {
+			filePath := utils.Concat(folderPath, "/main.yml")
+			os.Create(filePath)
+		}
+	}
 }
